Address_Geocoder
I needed to gather the latlong coordinates for quite a lot of addresses in order to use them for heatmaps and so on. The plan was to present the data in an existing app created in Oracle APEX, and I had no budget to use any pro-level geocoding services. This script in the result!
It queries the Oracle DB for addresses, queries Google Maps (using as many free API keys as you can throw at it), and inserts the latlong data.