Add files via upload
This is old and obsolete now. Jira has great boards built-in. But I wrote this back when I worked in a service desk and we needed something to display up on our wallboards. This one is a simple wallboard to connect to the Jira API and list tickets that haven't been updated since creation.